fix mod func
